requirements.txt
blinker==1.9.0
click==8.3.0
Flask==2.3.2
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysqlclient==2.2.0
packaging==25.0
pdfkit==1.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
